% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers_timestamp.R
\name{add_date}
\alias{add_date}
\title{Adds column 'date' to a dataframe containing a timestamp variable.}
\usage{
add_date(data, utc_col = character(1), tz = "UTC", unit = "s")
}
\arguments{
\item{data}{(`dataframe`). A dataframe containing a UTC timestamp variable (in seconds or milliseconds).}

\item{utc_col}{(`character(1)`). Column name of the UTC timestamp.}

\item{tz}{(`character(1)`). The timezone. Defaults to using 'UTC'. See \code{\link[lubridate]{as_datetime}}.}

\item{unit}{(`character(1)`). The unit of the timestamps. Can be 's' or 'ms'. Defaults to 's' (seconds).}
}
\value{
(`dataframe`) with added variable 'date'.
}
\description{
Adds column 'date' to a dataframe containing a timestamp variable.
}
\seealso{
Other helper functions: \code{\link{add_date_time}},
  \code{\link{add_time}}, \code{\link{add_weekday}}
}
\concept{helper functions}
