% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers_timestamp.R
\name{add_date_time}
\alias{add_date_time}
\title{Adds column 'date_time' to a dataframe containing a timestamp variable.}
\usage{
add_date_time(data, utc_col = character(1), tz = "UTC", unit = "s")
}
\arguments{
\item{data}{dataframe containing column named 'timestamp', which should be a Unix Timestamp (in seconds or milliseconds).}

\item{utc_col}{character. Column name of the UTC timestamp.}

\item{tz}{timezone. Defaults to using 'UTC'. See \code{\link[lubridate]{as_datetime}}.}

\item{unit}{unit of the timestamps. Can be 's' or 'ms'. Defaults to 's' (seconds).}
}
\value{
dataframe with added variable 'date_time'.
}
\description{
Adds column 'date_time' to a dataframe containing a timestamp variable.
}
\seealso{
Other helper functions: \code{\link{add_date}},
  \code{\link{add_time}}, \code{\link{add_weekday}},
  \code{\link{sliding_window}}
}
\concept{helper functions}
