% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers_timestamp.R
\name{add_time}
\alias{add_time}
\title{Adds column 'time' to a dataframe containing a timestamp variable.}
\usage{
add_time(data, utc_col = character(1), tz = "UTC", unit = "s")
}
\arguments{
\item{data}{dataframe containing a UTC timestamp variable (in seconds or milliseconds).}

\item{utc_col}{character. Column name of the UTC timestamp.}

\item{tz}{timezone. Defaults to using 'UTC'. See \code{\link[lubridate]{as_datetime}}.}

\item{unit}{unit of the timestamps. Can be 's' or 'ms'. Defaults to 's' (seconds).}
}
\value{
dataframe with added variable 'time'.
}
\description{
Adds column 'time' to a dataframe containing a timestamp variable.
}
\seealso{
Other helper functions: \code{\link{add_date_time}},
  \code{\link{add_date}}, \code{\link{add_weekday}}
}
\concept{helper functions}
