% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers_timestamp.R
\name{add_weekday}
\alias{add_weekday}
\title{Adds column 'weekday' to a dataframe containing a timestamp variable.}
\usage{
add_weekday(data, utc_col = character(1), tz = "UTC", unit = "s",
  week_start = 1, locale = "English_United States.1252")
}
\arguments{
\item{data}{dataframe containing a UTC timestamp variable (in seconds or milliseconds).}

\item{utc_col}{character. Column name of the UTC timestamp.}

\item{tz}{timezone. Defaults to using 'UTC'. See \code{\link[lubridate]{as_datetime}}.}

\item{unit}{unit of the timestamps. Can be 's' or 'ms'. Defaults to 's' (seconds).}

\item{week_start}{day on which week starts following ISO conventions - 1 means Monday, 7 means Sunday (default). See \code{\link[lubridate]{wday}}.}

\item{locale}{locale to use for day names. Defaults to 'English_United States.1252'. See \code{\link[lubridate]{wday}}.}
}
\value{
dataframe with added ordered factor variable 'weekday'.
}
\description{
Adds column 'weekday' to a dataframe containing a timestamp variable.
}
\seealso{
Other helper functions: \code{\link{add_date_time}},
  \code{\link{add_date}}, \code{\link{add_time}},
  \code{\link{sliding_window}}
}
\concept{helper functions}
