% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/batchtools.R
\name{addBatchtoolsAlgorithms}
\alias{addBatchtoolsAlgorithms}
\title{Adds batchtools algorithms and experiments.}
\usage{
addBatchtoolsAlgorithms(project)
}
\arguments{
\item{project}{fxtract_project object created by \code{\link{makeProject}}}
}
\description{
User defined functions must be saved as separate .R files in the projects subfolder 'feature_functions'.
See the tutorial for more information. These functions need a dataframe as input and a named vector as output.
}
\examples{
\dontrun{
addBatchtoolsAlgorithms(project = project)
}
}
