% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/batchtools.R
\name{addBatchtoolsProblems}
\alias{addBatchtoolsProblems}
\title{Adds batchtools problems.}
\usage{
addBatchtoolsProblems(project)
}
\arguments{
\item{project}{fxtract_project file created by \code{\link{makeProject}}}
}
\description{
Every grouping variable's dataframe is added as a batchtools problem.
}
\examples{
\dontrun{
addBatchtoolsProblems(project = project)
}
}
