% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addColumnByGroup.R
\name{addColumnByGroup}
\alias{addColumnByGroup}
\title{Helper function. Needs a function, which has a dataframe as input and a vector of length nrow(dataframe) as output.
This function will be evaluated grouped by \code{group_col}.}
\usage{
addColumnByGroup(data, group_col, fun, colname, check_fun = TRUE)
}
\arguments{
\item{data}{dataframe.}

\item{group_col}{character. Name of column, which contains identifiers on which the dataset should be grouped by. E.g. different user IDs.}

\item{fun}{function. Must be a function, which has a dataframe as input and a vector (of length nrow(dataframe)) as output.}

\item{colname}{character. Column name of the new column.}

\item{check_fun}{logical. If \code{TRUE}, fun(data) will be evaluated and checked, 
if the outcome is of correct form. Set to \code{FALSE}, if evaluation on the whole dataset takes too long.}
}
\value{
dataframe with added column, where a function \code{fun} is evaluated grouped by \code{group_col}.
}
\description{
Helper function. Needs a function, which has a dataframe as input and a vector of length nrow(dataframe) as output.
This function will be evaluated grouped by \code{group_col}.
}
\seealso{
Other helper functions: \code{\link{addColumn}},
  \code{\link{addDateTime}}, \code{\link{addDate}},
  \code{\link{addStudyDayByGroup}},
  \code{\link{addStudyDay}}, \code{\link{addTime}},
  \code{\link{addWeekday}}, \code{\link{calcStudyDay}},
  \code{\link{divideDataIntoIntervals}},
  \code{\link{slidingWindow2}}, \code{\link{slidingWindow}}
}
