% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addColumnByUserId.R
\name{addColumnByUserId}
\alias{addColumnByUserId}
\title{Helper function. Takes the mininum of timestamp as first day and labels oncoming days with the time difference to this day.}
\usage{
addColumnByUserId(data, fun)
}
\arguments{
\item{data}{dataframe containing column named 'userId', which should be the unique identifier for each ID.}

\item{fun}{function, which has a dataframe as input and a vector (of length nrow(dataframe)) as output.}
}
\value{
dataframe with added column
}
\description{
Helper function. Takes the mininum of timestamp as first day and labels oncoming days with the time difference to this day.
}
\seealso{
Other helper functions: \code{\link{addDateTime}},
  \code{\link{addDate}}, \code{\link{addStudyDay}},
  \code{\link{addTime}}, \code{\link{addWeekday}}
}
\concept{helper functions}
