% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpersTimestamp.R
\name{addDate}
\alias{addDate}
\title{Adds column 'date' to a dataframe containing a timestamp variable.}
\usage{
addDate(data, utc_col = character(1), tz = "UTC", unit = "s")
}
\arguments{
\item{data}{dataframe containing column named 'timestamp', which should be a Unix Timestamp (in seconds or milliseconds).}

\item{utc_col}{character. Column name of the UTC timestamp.}

\item{tz}{timezone. Defaults to using 'UTC'. See \code{\link[lubridate]{as_datetime}}.}

\item{unit}{unit of the timestamps. Can be 's' or 'ms'. Defaults to 's' (seconds).}
}
\value{
dataframe with added variable 'date'.
}
\description{
Adds column 'date' to a dataframe containing a timestamp variable.
}
\seealso{
Other helper functions: \code{\link{addDateTime}},
  \code{\link{addTime}}, \code{\link{addWeekday}},
  \code{\link{slidingWindow}}
}
\concept{helper functions}
