% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{addDateTime}
\alias{addDateTime}
\title{Adds column 'date_time' to a dataframe containing a timestamp variable.}
\usage{
addDateTime(data, tz = "UTC", unit = "s")
}
\arguments{
\item{data}{dataframe containing column named 'timestamp'.}

\item{tz}{timezone. Defaults to using 'UTC'. See \code{\link[lubridate]{as_datetime}}.}

\item{unit}{unit of timezone. Can be 's' or 'ms'. Defaults to 's' (seconds).}
}
\value{
dataframe with added variable 'date_time'.
}
\description{
Adds column 'date_time' to a dataframe containing a timestamp variable.
}
\seealso{
Other helper functions: \code{\link{addColumnByUserId}},
  \code{\link{addDate}}, \code{\link{addTime}},
  \code{\link{addWeekday}}, \code{\link{calcStudyDay}},
  \code{\link{slidingWindow}}
}
