% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{addStudyDay}
\alias{addStudyDay}
\title{Adds column 'studyDay' to a dataframe containing a timestamp variable.}
\usage{
addStudyDay(data)
}
\arguments{
\item{data}{dataframe containing column named 'timestamp'.}
}
\value{
dataframe with added ordered factor variable 'studyDay'. The first day for each userId will be 'day1'.
}
\description{
Adds column 'studyDay' to a dataframe containing a timestamp variable.
}
\seealso{
Other helper functions: \code{\link{addDateTime}},
  \code{\link{addDate}}, \code{\link{addTime}},
  \code{\link{addWeekday}}
}
\concept{helper functions}
