% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{addStudyDayByGroup}
\alias{addStudyDayByGroup}
\title{Helper function. For each group variable it takes the mininum of timestamp as first day and labels oncoming days with the time difference to this day.}
\usage{
addStudyDayByGroup(data, group_col, colname = "studyDay", ordered = TRUE)
}
\arguments{
\item{data}{dataframe.}

\item{group_col}{character. Name of column, which contains identifiers on which the dataset should be grouped by. E.g. different user IDs.}

\item{colname}{character. Column name of the new column.}

\item{ordered}{logical. If \code{TRUE}, the added character variable will be ordered, like 'day1' < 'day2' < 'day3' < ... .}
}
\value{
dataframe with added factor variable 'studyDay' for each group variable. The first day for each variable will be 'day1' .
}
\description{
Helper function. For each group variable it takes the mininum of timestamp as first day and labels oncoming days with the time difference to this day.
}
\seealso{
Other helper functions: \code{\link{addColumnByGroup}},
  \code{\link{addColumn}}, \code{\link{addDateTime}},
  \code{\link{addDate}}, \code{\link{addStudyDay}},
  \code{\link{addTime}}, \code{\link{addWeekday}},
  \code{\link{calcStudyDay}},
  \code{\link{divideDataIntoIntervals}},
  \code{\link{slidingWindow}}
}
