% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{addStudyDayPerUserId}
\alias{addStudyDayPerUserId}
\title{Helper function. For each userId it takes the mininum of timestamp as first day and labels oncoming days with the time difference to this day.}
\usage{
addStudyDayPerUserId(data, colname = "studyDay", ordered = TRUE)
}
\arguments{
\item{data}{dataframe. Dataframe containing column named 'userId', which should be the unique identifier for each ID.
Also a column named 'timestamp' is needed, which should be a Unix Timestamp (in seconds or milliseconds).}

\item{colname}{character. Column name of the new column.}

\item{ordered}{logical. If \code{TRUE}, the added character variable will be ordered, like 'day1' < 'day2' < 'day3' < ... .}
}
\value{
dataframe with added factor variable 'studyDay' for each userId. The first day for each userId will be 'day1' .
}
\description{
Helper function. For each userId it takes the mininum of timestamp as first day and labels oncoming days with the time difference to this day.
}
\seealso{
Other helper functions: \code{\link{addColumnByUserId}},
  \code{\link{addColumn}}, \code{\link{addDateTime}},
  \code{\link{addDate}}, \code{\link{addStudyDay}},
  \code{\link{addTime}}, \code{\link{addWeekday}},
  \code{\link{calcStudyDay}}, \code{\link{slidingWindow}}
}
