% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{addWeekday}
\alias{addWeekday}
\title{Adds column 'weekday' to a dataframe containing a timestamp variable.}
\usage{
addWeekday(data, tz = "UTC", unit = "s", week_start = 1,
  locale = "English_United States.1252")
}
\arguments{
\item{data}{dataframe containing column named 'timestamp'.}

\item{tz}{timezone. Defaults to using UTC.}

\item{unit}{unit of timezone. Can be 's' or 'ms'. Defaults to s (seconds).}

\item{week_start}{day on which week starts following ISO conventions - 1 means Monday, 7 means Sunday (default).}

\item{locale}{locale to use for day names. Default to 'English_United States.1252'.}
}
\value{
dataframe with added ordered factor variable 'weekday'.
}
\description{
Adds column 'weekday' to a dataframe containing a timestamp variable.
}
\seealso{
Other helper functions: \code{\link{addDateTime}},
  \code{\link{addDate}}, \code{\link{addStudyDay}},
  \code{\link{addTime}}
}
\concept{helper functions}
