% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcFeature.R
\name{calcFeature}
\alias{calcFeature}
\title{Calculates a given function grouped by \code{group_col}.}
\usage{
calcFeature(data, group_col, fun, check_fun = TRUE, summarize, colname)
}
\arguments{
\item{data}{dataframe.}

\item{group_col}{character. Name of column, which contains identifiers on which the dataset should be grouped by. E.g. different user IDs.}

\item{fun}{function. Must be a function, which has a dataframe as input and a (named) vector of desired length as output.}

\item{check_fun}{logical. If \code{TRUE}, fun(data) will be evaluated and checked, 
if the outcome is of correct form. Set to \code{FALSE}, if evaluation on the whole dataset takes too long.}

\item{summarize}{function. If more than one value is returned for each group, this function summarizes these values to one value or more values. E.g. mean or/and sd.}

\item{colname}{character. If one variable is returned for each group, you can specify a custom column name for this new column.}
}
\value{
dataframe
}
\description{
Calculates a given function grouped by \code{group_col}.
}
\examples{
fun = function(data) {
  c(uses_chrome = nrow(dplyr::filter(data, RUNNING_TASKS_baseActivity_mPackage == "com.android.chrome")))
}
calcFeature(data = studentlife_small, group_col = "userId", fun = fun)
}
\concept{feature functions}
