% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcFeatureByGroup.R
\name{calcFeatureByGroup}
\alias{calcFeatureByGroup}
\title{Calculates a given function grouped by \code{group_col}. For each group variable one single value is returned.}
\usage{
calcFeatureByGroup(data, group_col, fun, colname)
}
\arguments{
\item{data}{dataframe.}

\item{group_col}{character. Name of column, which contains identifiers on which the dataset should be grouped by. E.g. different user IDs.}

\item{fun}{function. Must be a function, which has a dataframe as input and a vector of length 1 as output.}

\item{colname}{character. Desired column name.}
}
\value{
dataframe with columns \code{group_col} and \code{colname}.
}
\description{
Calculates a given function grouped by \code{group_col}. For each group variable one single value is returned.
}
\examples{
fun = function(data) {
  nrow(dplyr::filter(data, RUNNING_TASKS_baseActivity_mPackage == "com.android.chrome"))
}
calcFeatureByGroup(data = studentlife.small, group_col = "userId", fun = fun,
  colname = "number_uses_chrome")
}
\seealso{
Other feature functions: \code{\link{calcFeaturePerStudyDayAndByGroup}}
}
