% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcFeaturePerStudyDayAndByGroup.R
\name{calcFeaturePerStudyDayAndByGroup}
\alias{calcFeaturePerStudyDayAndByGroup}
\title{Calculates a given function on the dataset for each studyDay and grouped by the column 'userId'.
Needs a function, which has a dataframe as input and a singular value as output.
A summary function, which summarizes the results for each day, will be evaluated for each 'userId'.}
\usage{
calcFeaturePerStudyDayAndByGroup(data, group_col, fun, colname, summary_fun,
  max_study_day, impute_empty_day = NA, export_results_per_day = FALSE)
}
\arguments{
\item{data}{dataframe.}

\item{group_col}{character. Name of column, which contains identifiers on which the dataset should be grouped by. E.g. different user IDs.}

\item{fun}{function. Must be a function, which has a dataframe as input and a vector (of length nrow(dataframe)) as output.}

\item{colname}{character. Column name of the new column.}

\item{summary_fun}{function. Must be a function, which has a vector as input and a singular value as output.}

\item{max_study_day}{integer. Function will be evaluated until the max study day.}

\item{impute_empty_day}{can be character, factor, or numerical.
Results on days, which did not have any data, will be imputed with this value.}

\item{export_results_per_day}{logical. If TRUE, an additional dataframe, where \code{fun} is evaluated at each studyDay, is returned.}
}
\value{
dataframe with columns userId and \code{colname}. One value for each userId.
}
\description{
Calculates a given function on the dataset for each studyDay and grouped by the column 'userId'.
Needs a function, which has a dataframe as input and a singular value as output.
A summary function, which summarizes the results for each day, will be evaluated for each 'userId'.
}
\examples{
fun = function(data) {
  nrow(dplyr::filter(data, RUNNING_TASKS_baseActivity_mPackage == "com.android.chrome"))
} #number uses google chrome
summary_fun = function(x) mean(x, na.rm = TRUE)
data = addStudyDay(studentlife.small)
calcFeaturePerStudyDayAndByGroup(data = data, group_col = "userId", fun = fun,
  colname = "mean_uses_chrome_per_day", summary_fun = summary_fun)
}
\seealso{
Other feature functions: \code{\link{calcFeatureByGroup}}
}
