% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcFeaturePerUserId.R
\name{calcFeaturePerUserId}
\alias{calcFeaturePerUserId}
\title{Calculates a given function for each userId}
\usage{
calcFeaturePerUserId(data, fun, colname)
}
\arguments{
\item{data}{dataframe. dataframe containing column named 'userId', which should be the unique identifier for each ID.}

\item{fun}{function. Must be a function, which has a dataframe as input and a vector of length 1 as output.}

\item{colname}{character. Desired column name.}
}
\value{
dataframe with columns userId and \code{colname}.
}
\description{
Calculates a given function for each userId
}
\examples{
myFun = function(data) {
  nrow(dplyr::filter(data, RUNNING_TASKS_baseActivity_mPackage == "com.android.chrome"))
}
calcFeaturePerUserId(data = studentlife.small, fun = myFun, colname = "number_uses_chrome")
}
