% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers_timestamp.R
\name{check_time_format}
\alias{check_time_format}
\title{Helper function. Checks if character has correct time format 'hh:mm:ss'.}
\usage{
check_time_format(time)
}
\arguments{
\item{time}{(`character()`). The character vector to be checked.}
}
\description{
Helper function. Checks if character has correct time format 'hh:mm:ss'.
}
