% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{checkCols}
\alias{checkCols}
\title{Helper function. Checks if data contains column names.}
\usage{
checkCols(colname, data)
}
\arguments{
\item{colname}{character. It will be checked, if the dataframe contains these column names.}

\item{data}{dataframe.}
}
\description{
Helper function. Checks if data contains column names.
}
