% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{checkTimeFormat}
\alias{checkTimeFormat}
\title{Helper function. Checks if character has correct time format 'hh:mm:ss'.}
\usage{
checkTimeFormat(x)
}
\arguments{
\item{x}{character vector.}
}
\description{
Helper function. Checks if character has correct time format 'hh:mm:ss'.
}
