% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/batchtools.R
\name{collectResults}
\alias{collectResults}
\title{Collects results.}
\usage{
collectResults(project)
}
\arguments{
\item{project}{fxtract_project object created by \code{\link{makeProject}}}
}
\value{
dataframe.
}
\description{
Summarizes all calculated features into one dataframe.
}
\examples{
\dontrun{
collectResults(project = project)
}
}
