% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/communicationNumberCalls.R
\name{communicationNumberCalls}
\alias{communicationNumberCalls}
\title{Calculates the total number of calls.}
\usage{
communicationNumberCalls(loggingdata, ids = unique(loggingdata$userId),
  call_type = "all", colname = "NumberCalls", progress_bar = FALSE)
}
\arguments{
\item{loggingdata}{Logging data from the Phonestudy LMU}

\item{ids}{A character vector of participant IDs}

\item{call_type}{\code{"all"} includes all kind of calls (OUTGOING, INCOMING, RINGING, MISSED, ONHOLD, and REJECTED).
Can also be \code{"OUTGOING"}, \code{"INCOMING"}, \code{"RINGING"}, \code{"MISSED"}, \code{"ONHOLD"}, or \code{"REJECTED"}.
Multiple types are accepted as input. In this case, the return value is the composite result for these types.}

\item{colname}{A character variable with the desired column name}

\item{progress_bar}{A logical variable. Set to TRUE, if a progress bar is wanted.}
}
\value{
Returns a dataframe with the total number of calls grouped by the participant IDs.
}
\description{
Calculates the total number of calls.
}
\examples{
communicationNumberCalls(phonedata, call_type = "INCOMING", colname = "number_inc_calls")
}
\concept{communication functions}
