% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/batchtools.R
\name{dataframeToRds}
\alias{dataframeToRds}
\title{Writes single RDS files from an R dataframe.}
\usage{
dataframeToRds(project, dataframe)
}
\arguments{
\item{project}{fxtract_project object created by \code{\link{makeProject}}}

\item{dataframe}{R dataframe.}
}
\value{
fxtract_project
}
\description{
Given an R dataframe with a grouping variable (defined by the project object), this function saves single RDS files for every grouping variable.
Parallelization is available via \link{foreach}.
}
\examples{
\dontrun{
dataframeToRds(project, studenlife.small)
}
}
