% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{divideDataIntoIntervals}
\alias{divideDataIntoIntervals}
\title{Helper function. Adds a new column, which divides the dataset into intervals of a given length.}
\usage{
divideDataIntoIntervals(data, interval, unit = "s", colname = "interval")
}
\arguments{
\item{data}{dataframe containing column named 'timestamp', which should be a Unix Timestamp (in seconds or milliseconds).}

\item{interval}{numeric. Number of seconds, which will be the length of the intervals.}

\item{unit}{unit of the timestamps. Can be 's' or 'ms'. Defaults to 's' (seconds).}

\item{colname}{character. Column name of the new column.}
}
\value{
character. This character divides the dataset into intervals.
}
\description{
Helper function. Adds a new column, which divides the dataset into intervals of a given length.
}
\seealso{
Other helper functions: \code{\link{addColumnByUserId}},
  \code{\link{addColumn}}, \code{\link{addDateTime}},
  \code{\link{addDate}},
  \code{\link{addStudyDayPerUserId}},
  \code{\link{addStudyDay}}, \code{\link{addTime}},
  \code{\link{addWeekday}}, \code{\link{calcStudyDay}},
  \code{\link{slidingWindow}}
}
