% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filterWeekday.R
\name{filterWeekday}
\alias{filterWeekday}
\title{Filter function. Filters the data by weekdays. The dataset can e.g. be filtered by all days between monday and friday.
The dataset needs an ordered factor variable named 'weekday'.}
\usage{
filterWeekday(data, from_day = "Mon", from_time = "00:00:00",
  until_day = "Sun", until_time = "23:59:59")
}
\arguments{
\item{data}{dataframe. dataframe containing column named 'userId', which should be the unique identifier for each ID.}

\item{from_day}{character. Day of the week where filtering should start. Please check the machines locale for correct usage. Defaults to "Mon".}

\item{from_time}{character. Time where filtering should start (on the day from_day). Must be a string with the format "hh:mm:ss". Defaults to "00:00:00".}

\item{until_day}{character. Day of the week where filtering should end. Please check the machines locale for correct usage. Defaults to "Sun".}

\item{until_time}{character. Time where filtering should end (on the day until_day). Must be a string with the format "hh:mm:ss". Defaults to "23:59:59".}
}
\value{
filtered dataframe by weekday and time.
}
\description{
Filter function. Filters the data by weekdays. The dataset can e.g. be filtered by all days between monday and friday.
The dataset needs an ordered factor variable named 'weekday'.
}
