% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/batchtools.R
\name{getProjectStatus}
\alias{getProjectStatus}
\title{Summarizes the projects status.}
\usage{
getProjectStatus(project)
}
\arguments{
\item{project}{fxtract_project object created by \code{\link{makeProject}}}
}
\value{
list.
}
\description{
A detailed summary of which feature functions were already calculated.
}
\examples{
\dontrun{
getProjectStatus(project = project)
}
}
