% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{lakers.ts}
\alias{lakers.ts}
\title{lakers}
\format{An object of class \code{data.frame} with 34624 rows and 15 columns.}
\usage{
lakers.ts
}
\description{
lakers dataset from the lubridate package: \link[lubridate]{lakers}. See also:
}
\details{
https://www.rdocumentation.org/packages/lubridate/versions/1.7.4/topics/lakers

A new column 'timestamp' was added for easier usage of this package. Also a new column 'userId', which indicates different games, was added.
Please note, that the timezone was set to 'America/Los_Angeles', ignoring different timezones.
}
\keyword{data}
