% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/batchtools.R
\name{makeProject}
\alias{makeProject}
\title{Iniates batchtools functionality}
\usage{
makeProject(newProjectName, ...)
}
\arguments{
\item{newProjectName}{character. The new project name. A subfolder under the (created) folder `projects` will be created.}

\item{...}{other arguments passed to \code{batchtools::makeExperimentRegistry}}
}
\value{
fxtract_project object.
}
\description{
This function iniates batchtools functionality of this package.
Check the tutorial on the github page!
Every desired function will be calculated for each member
of a grouping variable (e.g. each participant of a study).
The data will be saved as an RDS-file for each member of a grouping variable.
This enables the possibility to calculate features even if all raw data cannot be loaded into RAM.
Feature extraction can be paused and restarted at a later time, even on a different machine.
Parallelization can be done via \cite{batchtools}.
}
\examples{
\dontrun{
project = makeProject("my_project")
}
}
