% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{phonedata}
\alias{phonedata}
\title{phonedata}
\format{An object of class \code{data.frame} with 40 rows and 10 columns.}
\usage{
phonedata
}
\description{
Contains exemplary raw phone logs.
}
\keyword{data}
