% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/project.R
\name{Project}
\alias{Project}
\title{R6 Object for Feature Extraction.}
\format{\code{\link{R6Class}} object.}
\description{
\code{Project} calculates features from longitudinal data for each grouping variable individually with batchtools.
}
\section{Usage}{

\preformatted{
my_project = Project$new("my_project")
}
}

\section{Arguments}{


For Project$new():
\describe{
\item{project_name: }{('character(1)'): A user defined name of the project. All necessary data will be saved on the path: ./projects/project_name/}
\item{load: }{(`logical(1)`): If TRUE, an existing project will be loaded.}
}
}

\section{Details}{

All datasets and feature functions are saved in this R6 object. \code{Project} heavily relies on the R-package batchtools.
Data will be saved as single RDS files (for each grouping variable) and feature functions are calculated on each single dataset.
A big advantage of this method is that it scales nicely for larger datasets. Data is only loaded into RAM, when needed.
}

\section{Fields}{

\describe{
\item{project_name: }{(`character(1)`): The projects name.}
\item{dir: }{(`character(1)`): The projects directory.}
\item{group_by: }{(`character(1)`): The column on which to group by.}
\item{reg: }{(`registry`): batchtools registry.}
}
}

\section{Methods}{

\describe{
\item{add_data(dataframe, group_by)}{[dataframe: (`data.frame`)] A dataframe which shall be added to the R6 object. \cr
 [group_by: (`character(1)`)] The grouping variable of the dataframe. \cr \cr
 This method writes single RDS files (this can be parallelized with foreach) and adds them as batchtools problems for each grouping variable of a dataframe.
 After that, batchtools experiments will be added too.}
\item{preprocess_data(fun)}{[fun: (`function`)] A function, which has a dataframe as input and a dataframe as output. \cr \cr
 This method loads the RDS files and applies this function on them. The old RDS files are overwritten and the batchtools
 problems and experiments are updated.}
\item{remove_data(data)}{[data: (`character(1)`)] The grouping variable's name. \cr \cr
 This method deletes the RDS file and batchtools problem of one single grouping variable.}
\item{add_feature(fun)}{[fun: (`function`)] A function, which has a dataframe as input and a named vector as output. \cr \cr
 This method adds a function as batchtools algorithm. After that, batchtools experiments are added too.}
\item{remove_feature(fun)}{[fun: (`function | character(1)`)] A function (or the name of the function as character) which shall be removed. \cr \cr
 This method removes the batchtools algorithms and experiments corresponding to the given function.}
\item{calc_features()}{This method calculates all features on all datasets. Internally, it submits all batchtools jobs, which are not done yet.}
\item{get_project_status()}{This method gives an overview over which features are done on which datasets.}
\item{collect_results()}{This method returns a dataframe with the calculated features.}
\item{plot()}{[internal] method to print the R6 object.}
\item{\code{clone()}}{[internal] method to clone the R6 object.}
\item{\code{initialize()}}{[internal] method to initialize the R6 object.}
}
}

\examples{
unlink("projects/my_project", recursive = TRUE)
my_project = Project$new("my_project")
my_project$add_data(iris, group_by = "Species")
fun = function(data) {
  c(mean_sepal_length = mean(data$Sepal.Length))
}
my_project$add_feature(fun)
my_project$calc_features()
my_project$collect_results()
unlink("projects/my_project", recursive = TRUE)
}
