% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/slidingWindow.R
\name{slidingWindow}
\alias{slidingWindow}
\title{FIXME: add documentation}
\usage{
slidingWindow(data, fun, steps, time_in_seconds, colname)
}
\arguments{
\item{data}{dataframe. dataframe containing column named 'userId', which should be the unique identifier for each ID.}

\item{fun}{function. Must be a function, which has a dataframe as input and a vector of length 1 as output.}

\item{steps}{integer. Number of steps which the sliding window should go back.}

\item{time_in_seconds}{integer. Number of seconds which the sliding window should go back.}

\item{colname}{character. Column name of the new column.}
}
\value{
dataframe with added column
}
\description{
FIXME: add documentation
}
\seealso{
Other helper functions: \code{\link{addColumnByUserId}},
  \code{\link{addDateTime}}, \code{\link{addDate}},
  \code{\link{addTime}}, \code{\link{addWeekday}},
  \code{\link{calcStudyDay}}
}
