% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/batchtools.R
\name{sqlToRds}
\alias{sqlToRds}
\title{Writes single RDS files from an SQL database.}
\usage{
sqlToRds(project, file.dir, tbl_name)
}
\arguments{
\item{project}{fxtract_project object created by \code{\link{makeProject}}}

\item{file.dir}{character. The file directory of the SQL database.}

\item{tbl_name}{character. The table name required by \code{\link[dplyr]{tbl}}.}
}
\value{
fxtract_project
}
\description{
Given an SQL database with a grouping variable, this function loads the SQL database into RAM for each grouping variable
and saves these as dataframes in RDS files. Parallelization is available via \link{foreach}.
}
\examples{
\dontrun{
project = sqlToRds(project, "projects/myProject/SQL_database.sql",
  tbl_name = "table")
}
}
