% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{studentlife_small}
\alias{studentlife_small}
\title{studentlife_small}
\format{An object of class \code{data.frame} with 162154 rows and 19 columns.}
\usage{
studentlife_small
}
\description{
Contains exemplary gps and app data from the studentlife dataset (users: u00, u01, u02)
}
\details{
http://studentlife.cs.dartmouth.edu/dataset.html

Wang, Rui, Fanglin Chen, Zhenyu Chen, Tianxing Li, Gabriella Harari, Stefanie Tignor, Xia Zhou, Dror Ben-Zeev, and Andrew T. Campbell. "StudentLife: Assessing Mental Health, Academic Performance and Behavioral Trends of College Students using Smartphones." In Proceedings of the ACM Conference on Ubiquitous Computing. 2014.
}
\keyword{data}
