% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers_timestamp.R
\name{time_to_sec}
\alias{time_to_sec}
\title{Helper function. Converts "hh:mm:ss" into number of seconds.}
\usage{
time_to_sec(x)
}
\arguments{
\item{x}{character vector.}
}
\value{
numeric vector. Number of seconds.
}
\description{
Helper function. Converts "hh:mm:ss" into number of seconds.
}
