% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{timeToSec}
\alias{timeToSec}
\title{Helper function. Converts "hh:mm:ss" into number of seconds.}
\usage{
timeToSec(x)
}
\arguments{
\item{x}{character vector.}
}
\value{
numeric. Number of seconds.
}
\description{
Helper function. Converts "hh:mm:ss" into number of seconds.
}
