% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/batchtools.R
\name{useDataframe}
\alias{useDataframe}
\title{Writes single RDS files from an R dataframe.}
\usage{
useDataframe(project, dataframe, group_by)
}
\arguments{
\item{project}{fxtract_project object created by \code{\link{makeProject}}.}

\item{dataframe}{R dataframe.}

\item{group_by}{character. The grouping variable.}
}
\value{
fxtract_project
}
\description{
Given an R dataframe with a grouping variable (defined by the project object), this function saves single RDS files for every grouping variable.
Parallelization is available via \link{foreach}.
}
\examples{
\dontrun{
useDataframe(project, studenlife.small)
}
}
