% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addColumn.R
\name{addColumn}
\alias{addColumn}
\title{Helper function. Needs a function, which has a dataframe as input and a vector of length nrow(dataframe) as output.}
\usage{
addColumn(data, fun, colname)
}
\arguments{
\item{data}{dataframe.}

\item{fun}{function. Must be a function, which has a dataframe as input and a vector (of length nrow(dataframe)) as output.}

\item{colname}{character. Column name of the new column.}
}
\value{
dataframe with added column.
}
\description{
Helper function. Needs a function, which has a dataframe as input and a vector of length nrow(dataframe) as output.
}
\seealso{
Other helper functions: \code{\link{addColumnByGroup}},
  \code{\link{addDateTime}}, \code{\link{addDate}},
  \code{\link{addStudyDayByGroup}},
  \code{\link{addStudyDay}}, \code{\link{addTime}},
  \code{\link{addWeekday}}, \code{\link{calcStudyDay}},
  \code{\link{divideDataIntoIntervals}},
  \code{\link{slidingWindow}}
}
