% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{addDate}
\alias{addDate}
\title{Adds column 'date' to a dataframe containing a timestamp variable.}
\usage{
addDate(data, tz = "UTC", unit = "s")
}
\arguments{
\item{data}{dataframe containing column named 'timestamp'.}

\item{tz}{timezone. Defaults to using 'UTC'. See \code{\link[lubridate]{as_datetime}}.}

\item{unit}{unit of timezone. Can be 's' or 'ms'. Defaults to s (seconds).}
}
\value{
dataframe with added variable 'date'.
}
\description{
Adds column 'date' to a dataframe containing a timestamp variable.
}
\seealso{
Other helper functions: \code{\link{addColumnByUserId}},
  \code{\link{addDateTime}}, \code{\link{addTime}},
  \code{\link{addWeekday}}, \code{\link{calcStudyDay}}
}
