% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpersStudyDay.R
\name{addStudyDay}
\alias{addStudyDay}
\title{Helper function. It takes the mininum of timestamp as first day and labels oncoming days with the time difference to this day.}
\usage{
addStudyDay(data, colname = "studyDay", ordered = TRUE)
}
\arguments{
\item{data}{dataframe containing column named 'timestamp', which should be a Unix Timestamp (in seconds or milliseconds).}

\item{colname}{character. Column name of the new column.}

\item{ordered}{logical. If \code{TRUE}, the added factor variable will be ordered, like 'day1' < 'day2' < 'day3' < ... .}
}
\value{
dataframe with added factor variable 'studyDay'. The first day will be 'day1' .
}
\description{
Helper function. It takes the mininum of timestamp as first day and labels oncoming days with the time difference to this day.
}
\seealso{
Other helper functions: \code{\link{addColumnByGroup}},
  \code{\link{addColumn}}, \code{\link{addDateTime}},
  \code{\link{addDate}}, \code{\link{addStudyDayByGroup}},
  \code{\link{addTime}}, \code{\link{addWeekday}},
  \code{\link{calcStudyDay}},
  \code{\link{divideDataIntoIntervals}},
  \code{\link{slidingWindow}}
}
