% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{addTime}
\alias{addTime}
\title{Adds column 'time' to a dataframe containing a timestamp variable.}
\usage{
addTime(data, tz = "UTC", unit = "s")
}
\arguments{
\item{data}{dataframe containing column named 'timestamp'.}

\item{tz}{timezone. Defaults to using UTC.}

\item{unit}{unit of timezone. Can be 's' or 'ms'. Defaults to s (seconds).}
}
\value{
dataframe with added variable 'time'.
}
\description{
Adds column 'time' to a dataframe containing a timestamp variable.
}
\seealso{
Other helper functions: \code{\link{addDateTime}},
  \code{\link{addDate}}, \code{\link{addStudyDay}},
  \code{\link{addWeekday}}
}
\concept{helper functions}
