% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{addWeekday}
\alias{addWeekday}
\title{Adds column 'weekday' to a dataframe containing a timestamp variable.}
\usage{
addWeekday(data, tz = "UTC", unit = "s", week_start = 1,
  locale = "English_United States.1252")
}
\arguments{
\item{data}{dataframe containing column named 'timestamp', which should be a Unix Timestamp (in seconds or milliseconds).}

\item{tz}{timezone. Defaults to using 'UTC'. See \code{\link[lubridate]{as_datetime}}.}

\item{unit}{unit of the timestamps. Can be 's' or 'ms'. Defaults to 's' (seconds).}

\item{week_start}{day on which week starts following ISO conventions - 1 means Monday, 7 means Sunday (default). See \code{\link[lubridate]{wday}}.}

\item{locale}{locale to use for day names. Defaults to 'English_United States.1252'. See \code{\link[lubridate]{wday}}.}
}
\value{
dataframe with added ordered factor variable 'weekday'.
}
\description{
Adds column 'weekday' to a dataframe containing a timestamp variable.
}
\seealso{
Other helper functions: \code{\link{addColumnByUserId}},
  \code{\link{addColumn}}, \code{\link{addDateTime}},
  \code{\link{addDate}},
  \code{\link{addStudyDayPerUserId}},
  \code{\link{addStudyDay}}, \code{\link{addTime}},
  \code{\link{calcStudyDay}},
  \code{\link{divideDataIntoIntervals}},
  \code{\link{slidingWindow}}
}
