% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcFeature.R
\name{calcFeature}
\alias{calcFeature}
\title{Calculates a given function grouped by \code{group_col}.}
\usage{
calcFeature(data, group_col, fun, check_fun = TRUE)
}
\arguments{
\item{data}{dataframe.}

\item{group_col}{character. Name of column, which contains identifiers on which the dataset should be grouped by. E.g. different user IDs.}

\item{fun}{function. Must be a function, which has a dataframe as input and a vector of length 1 as output.}

\item{check_fun}{logical. If \code{TRUE}, fun(data) will be evaluated and checked, 
if the outcome is of correct form. Set to \code{FALSE}, if evaluation on the whole dataset takes too long.}
}
\value{
dataframe
}
\description{
Calculates a given function grouped by \code{group_col}.
}
\examples{
number_uses_chrome = function(data) {
  nrow(dplyr::filter(data, RUNNING_TASKS_baseActivity_mPackage == "com.android.chrome"))
}
calcFeature(data = studentlife.small, group_col = "userId", fun = number_uses_chrome)
}
\concept{feature functions}
