% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{calcStudyDay}
\alias{calcStudyDay}
\title{Helper function. Takes the mininum of timestamp as first day and labels oncoming days with the time difference to this day.}
\usage{
calcStudyDay(data)
}
\arguments{
\item{data}{dataframe containing column named 'timestamp'.}
}
\value{
dataframe with added ordered factor variable 'studyDay'. The first day for will be 'day1'.
}
\description{
Helper function. Takes the mininum of timestamp as first day and labels oncoming days with the time difference to this day.
}
\seealso{
Other helper functions: \code{\link{addColumnByUserId}},
  \code{\link{addDateTime}}, \code{\link{addDate}},
  \code{\link{addTime}}, \code{\link{addWeekday}}
}
