% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{divideDataIntoIntervals}
\alias{divideDataIntoIntervals}
\title{Helper function. Adds a new column, which divides the dataset into intervals of a given length.
The base R function \code{cut} is used with \code{include.lowest = TRUE}.
Values larger than the biggest cut point will be assigned a new interval.}
\usage{
divideDataIntoIntervals(data, steps, time_in_sec, unit = "s")
}
\arguments{
\item{data}{dataframe.}

\item{steps}{integer. Number of observations, which will be in the intervals.
Use this for data with fixed sample rate. \code{unit} will have no effect, if \code{steps} are chosen.}

\item{time_in_sec}{numeric. Number of seconds, which will be the length of the intervals.
Use this for data with variable sample rate.}

\item{unit}{unit of the timestamps. Can be 's' or 'ms'. Defaults to 's' (seconds).}
}
\value{
character. This character divides the dataset into intervals.
}
\description{
Helper function. Adds a new column, which divides the dataset into intervals of a given length.
The base R function \code{cut} is used with \code{include.lowest = TRUE}.
Values larger than the biggest cut point will be assigned a new interval.
}
\seealso{
Other helper functions: \code{\link{addColumnByGroup}},
  \code{\link{addColumn}}, \code{\link{addDateTime}},
  \code{\link{addDate}}, \code{\link{addStudyDayByGroup}},
  \code{\link{addStudyDay}}, \code{\link{addTime}},
  \code{\link{addWeekday}}, \code{\link{calcStudyDay}},
  \code{\link{slidingWindow}}
}
