% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers_timestamp.R
\name{filter_daytime}
\alias{filter_daytime}
\title{Filter function. Filters the data by daytime. The dataset can e.g. be filtered from "06:00:00" until "09:00:00".
The dataset needs an ordered factor variable named 'time' (best added by the function add_time).}
\usage{
filter_daytime(data, from_time = "07:00:00", until_time = "18:00:00")
}
\arguments{
\item{data}{(`dataframe`). Dataframe with a column named 'weekday'.}

\item{from_time}{(`character(1)`). Time where filtering should start (on the day from_day). Must be a string with the format "hh:mm:ss". Defaults to "00:00:00".}

\item{until_time}{(`character(1)`). Time where filtering should end (on the day until_day). Must be a string with the format "hh:mm:ss". Defaults to "23:59:59".}
}
\value{
The dataset filtered according to the given time constraints.
}
\description{
Filter function. Filters the data by daytime. The dataset can e.g. be filtered from "06:00:00" until "09:00:00".
The dataset needs an ordered factor variable named 'time' (best added by the function add_time).
}
\seealso{
Other filter functions: \code{\link{filter_weekday}}
}
\concept{filter functions}
