% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/batchtools.R
\name{loadProject}
\alias{loadProject}
\title{Loads fxtract project.}
\usage{
loadProject(file.dir, group_by)
}
\arguments{
\item{file.dir}{character. The project's directory.}

\item{group_by}{character. The grouping variable needs to be defined again.}
}
\value{
fxtract_project object
}
\description{
In order to restart feature extraction, the fxtract project needs to be loaded with this function.
}
\examples{
\dontrun{
project = loadProject("Projects/my_project", group_by = "user")
}
}
