% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/slidingWindow.R
\name{slidingWindow}
\alias{slidingWindow}
\title{Adds a variable (or a set of variables) to a dataset by a sliding window approach.}
\usage{
slidingWindow(data, fun, steps, time_in_sec, check_fun = TRUE,
  eval_at_rows = numeric(0), utc_col = character(1), unit = "s")
}
\arguments{
\item{data}{dataframe.}

\item{fun}{function. A function that has `data`` as input and a dataframe with 1 row as output.}

\item{steps}{integer. Number of steps which the sliding window should go back.
Use this for data with fixed sample rate.}

\item{time_in_sec}{integer. Number of seconds which the sliding window should go back.
Use this for data with variable sample rate.}

\item{check_fun}{logical. If \code{TRUE}, fun(data) will be evaluated and checked, 
if the outcome is of correct form. Set to \code{FALSE}, if evaluation on the whole dataset takes too long.}

\item{eval_at_rows}{integer. Integer vector at which rows the function should be executed. Defaults to `1:nrow(data)`.}

\item{utc_col}{character. Column name of the UTC timestamp.}

\item{unit}{unit of the timestamps. Can be 's' or 'ms'. Defaults to 's' (seconds).}
}
\value{
dataframe with added column(s).
}
\description{
Adds a variable (or a set of variables) to a dataset by a sliding window approach.
}
\seealso{
Other helper functions: \code{\link{addColumnByGroup}},
  \code{\link{addColumn}}, \code{\link{addDateTime}},
  \code{\link{addDate}}, \code{\link{addStudyDayByGroup}},
  \code{\link{addStudyDay}}, \code{\link{addTime}},
  \code{\link{addWeekday}}, \code{\link{calcStudyDay}},
  \code{\link{divideDataIntoIntervals}}
}
\concept{helper functions}
