% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/slidingWindow.R
\name{slidingWindow}
\alias{slidingWindow}
\title{Adds a variable (or a set of variables) to a dataset by a sliding window approach.}
\usage{
slidingWindow(data, fun, steps, time_in_sec, check_fun = TRUE)
}
\arguments{
\item{data}{dataframe.}

\item{fun}{function. A function that has `data`` as input and a dataframe with 1 row as output.}

\item{steps}{integer. Number of steps which the sliding window should go back.
Use this for data with fixed sample rate.}

\item{time_in_sec}{integer. Number of seconds which the sliding window should go back.
Use this for data with variable sample rate.}

\item{check_fun}{logical. If \code{TRUE}, fun(data) will be evaluated and checked, 
if the outcome is of correct form. Set to \code{FALSE}, if evaluation on the whole dataset takes too long.}
}
\value{
dataframe with added column(s).
}
\description{
Adds a variable (or a set of variables) to a dataset by a sliding window approach.
}
\examples{
fun = function(x) data.frame(mean.accuracy.last30 = mean(x$accuracy, na.rm = TRUE), 
  max.accuracy.last30 = max(x$accuracy, na.rm = TRUE))
data = addDateTime(studentlife.small[1:30, ]) #needs date_time variable
slidingWindow(data, fun = fun, time_in_sec = 60 * 60) #mean accuracy of last hour
}
\seealso{
Other helper functions: \code{\link{addColumnByGroup}},
  \code{\link{addColumn}}, \code{\link{addDateTime}},
  \code{\link{addDate}}, \code{\link{addStudyDayByGroup}},
  \code{\link{addStudyDay}}, \code{\link{addTime}},
  \code{\link{addWeekday}}, \code{\link{calcStudyDay}},
  \code{\link{divideDataIntoIntervals}}
}
